requirements.txt
Flask==2.1.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
Jinja2==3.1.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Mako==1.2.0
matplotlib==3.5.3
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy~=1.23.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
requests==2.27.1
requests-oauthlib==1.3.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
Werkzeug==2.1.1
wrapt==1.14.1
